cpu/msp430f5xyz: Added support for MSP430F5XXX and MSP430F6XXX CPUs
Based on the structure of cpu/msp430fxyz, support was added for the 5XXX
and 6XXX families of CPUs. Some of the changes present in the peripheral
files might be idiosyncratic to the board that I am using and will be
fixed in later commits.

cpu/msp430_common: Added support for MSP430X

Changes were made to msp430_common/cpu.c, msp430_common/include/cpu.h
and its Makefile in order to accomodate the extension for MSP430 that
allows for use of greater quantities of memory in available CPUs.